+++ api/db/database-builder/factory/build-user.js
@@ -2,9 +2,9 @@ import lodash from 'lodash';
 
 const { isUndefined, isNil } = lodash;
 
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../lib/domain/constants/identity-providers.js';
 import { AuthenticationMethod, Membership } from '../../../lib/domain/models/index.js';
 import { PIX_ADMIN } from '../../../src/authorization/domain/constants.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { cryptoService } from '../../../src/shared/domain/services/crypto-service.js';
 import { databaseBuffer } from '../database-buffer.js';
 import { buildCertificationCenter } from './build-certification-center.js';
