+++ api/lib/domain/usecases/remove-authentication-method.js
@@ -1,4 +1,4 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../constants/identity-providers.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
 import * as OidcIdentityProviders from '../constants/oidc-identity-providers.js';
 import { UserNotAuthorizedToRemoveAuthenticationMethod } from '../errors.js';
