+++ api/src/shared/domain/services/user-service.js
@@ -1,6 +1,6 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
 import { AuthenticationMethod } from '../../../../lib/domain/models/AuthenticationMethod.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../identity-access-management/domain/constants/identity-providers.js';
 import { UserToCreate } from '../../../identity-access-management/domain/models/UserToCreate.js';
 
 /**
