+++ api/src/shared/infrastructure/repositories/authentication-method-repository.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 
 import { knex } from '../../../../db/knex-database-connection.js';
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
 import * as OidcIdentityProviders from '../../../../lib/domain/constants/oidc-identity-providers.js';
 import { AlreadyExistingEntityError, AuthenticationMethodNotFoundError } from '../../../../lib/domain/errors.js';
 import { AuthenticationMethod } from '../../../../lib/domain/models/AuthenticationMethod.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../identity-access-management/domain/constants/identity-providers.js';
 import * as knexUtils from '../utils/knex-utils.js';
 
 const AUTHENTICATION_METHODS_TABLE = 'authentication-methods';
