+++ api/tests/acceptance/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -1,4 +1,4 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import {
   createServer,
   databaseBuilder,
