+++ api/tests/acceptance/application/users/get-user-authentication-methods-route-get_test.js
@@ -1,4 +1,4 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import {
   createServer,
   databaseBuilder,
