+++ api/tests/shared/integration/infrastructure/repositories/user-repository_test.js
@@ -1,6 +1,5 @@
 import lodash from 'lodash';
 const { each, map, times, pick } = lodash;
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../lib/domain/constants/identity-providers.js';
 import * as OidcIdentityProviders from '../../../../../lib/domain/constants/oidc-identity-providers.js';
 import {
   AlreadyExistingEntityError,
@@ -15,6 +14,7 @@ import { Membership } from '../../../../../lib/domain/models/Membership.js';
 import { UserDetailsForAdmin } from '../../../../../lib/domain/models/UserDetailsForAdmin.js';
 import { OrganizationLearnerForAdmin } from '../../../../../lib/domain/read-models/OrganizationLearnerForAdmin.js';
 import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { Organization } from '../../../../../src/organizational-entities/domain/models/Organization.js';
 import { User } from '../../../../../src/shared/domain/models/User.js';
 import * as userRepository from '../../../../../src/shared/infrastructure/repositories/user-repository.js';
