+++ api/tests/shared/prescriber-management/unit/infrastructure/serializers/jsonapi/prescriber-serializer_test.js
@@ -1,5 +1,5 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../../../lib/domain/constants/identity-providers.js';
 import { Membership } from '../../../../../../../lib/domain/models/index.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { ORGANIZATION_FEATURE } from '../../../../../../../src/shared/domain/constants.js';
 import * as serializer from '../../../../../../../src/shared/prescriber-management/infrastructure/serializers/jsonapi/prescriber-serializer.js';
 import { domainBuilder, expect } from '../../../../../../test-helper.js';
