+++ api/tests/unit/application/users/user-controller_test.js
@@ -1,5 +1,4 @@
 import { userController } from '../../../../lib/application/users/user-controller.js';
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
 import { eventBus } from '../../../../lib/domain/events/index.js';
 import { UserAnonymized } from '../../../../lib/domain/events/UserAnonymized.js';
 import { UserOrganizationForAdmin } from '../../../../lib/domain/read-models/UserOrganizationForAdmin.js';
@@ -8,6 +7,7 @@ import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransact
 import * as requestResponseUtils from '../../../../lib/infrastructure/utils/request-response-utils.js';
 import { usecases as devcompUsecases } from '../../../../src/devcomp/domain/usecases/index.js';
 import { evaluationUsecases } from '../../../../src/evaluation/domain/usecases/index.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { User } from '../../../../src/shared/domain/models/User.js';
 import * as queryParamsUtils from '../../../../src/shared/infrastructure/utils/query-params-utils.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
