+++ api/tests/unit/domain/models/AuthenticationMethod_test.js
@@ -1,7 +1,7 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
 import * as OidcIdentityProviders from '../../../../lib/domain/constants/oidc-identity-providers.js';
 import { ObjectValidationError } from '../../../../lib/domain/errors.js';
 import { AuthenticationMethod } from '../../../../lib/domain/models/AuthenticationMethod.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { expect } from '../../../test-helper.js';
 
 describe('Unit | Domain | Models | AuthenticationMethod', function () {
