+++ api/tests/unit/domain/usecases/update-expired-password_test.js
@@ -1,6 +1,6 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../lib/domain/constants/identity-providers.js';
 import { UserNotFoundError } from '../../../../lib/domain/errors.js';
 import { updateExpiredPassword } from '../../../../lib/domain/usecases/update-expired-password.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../src/identity-access-management/domain/constants/identity-providers.js';
 import { ForbiddenAccess } from '../../../../src/shared/domain/errors.js';
 import { logger } from '../../../../src/shared/infrastructure/utils/logger.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
